fixed issue 19
