+++ .release-it.json
@@ -3,7 +3,7 @@
     "before:init": ["npm run lint", "npm test"],
     "after:git:release": "echo After git push, before github release",
     "after:release": "echo Successfully released ${name} v${version} to ${repo.repository}.",
-    "after:bump": "npx auto-changelog"
+    "after:bump": ["npm run docs", "npx auto-changelog"]
   },
   "git": {
     "changelog": "npx auto-changelog",
