+++ lib/index.js
@@ -3,11 +3,11 @@
  * @module stonksjs
  */
 const screeners = require('./screeners');
-const robinhood = require('./robinhood');
+const Robinhood = require('./robinhood');
 const scrapers = require('./scrapers');
 
 module.exports = {
   screeners,
-  robinhood,
   scrapers,
+  Robinhood,
 };
